requirements.txt
requests==2.32.5
urllib3==2.6.3
discord.py==2.6.4
asyncio==4.0.0